Add files via upload
This update allows for the functionality of Attack and AttackResult to have "attached" versions. This simulates multiple attacks happening "at the same time." The conditions for attacks are used appropriately. This also includes the addition of Conditions being attached to attacks as single-use, for that attack only, conditions that affect the attacker or defender as specified. As of this upload this did not appear to add any additional bugs, but testing has not been completed.