Add files via upload
Restore Health function in Character altered to display the results rather than return the integer of amount healed. Other classes edited with respect to this change. Additionally, lifesteal functionality was added to attacks, so a specified double percentage of lifesteal can be specified, and the attack will one-time heal from that attack before any Conditions on the attack are removed.