Add files via upload
The Sentinel Specialist has been added with the new format with all the Abilities created as separate files for now (for easier filling and reference). It is now possible to create Sentinel Specialist objects without errors with correct base stats, but none of the abilities have been implemented past their initialization stage.