Add files via upload
Multi-Purposed and Flaming Arrow completely finished with correct functionality. In addition, the normal execute in Attack was edited to return the attack result made so that abilities can have additional effects without relying on "Previous Attack" which may be inconsistent due to extra attacks happening in pre/post attack effects.